fix: overflow recursion limit bug

+++ src/styles/constants.rs
@@ -33,8 +33,8 @@ pub const ICON_FONT_SIZE_TOOLBAR: u16 = 24;
 
 pub const SIDEBAR_WIDTH: u16 = 200;
 
-pub const WINDOW_INITIAL_WIDTH: u32 = 1028;
-pub const WINDOW_INITIAL_HEIGHT: u32 = 640;
+pub const WINDOW_INITIAL_WIDTH: f32 = 1028.0;
+pub const WINDOW_INITIAL_HEIGHT: f32 = 640.0;
 
 pub const WINDOW_MIN_WIDTH: u32 = 800;
 pub const WINDOW_MIN_HEIGHT: u32 = 450;
